backend/pyproject.toml
[tool.poetry]
name = "honda-backend"
version = "0.1.0"
description = "Honda internal system backend"
authors = ["Honda Team <team@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.2"
uvicorn = { extras = ["standard"], version = "^0.30.6" }
sqlalchemy = "^2.0.37"
asyncpg = "^0.29.0"
alembic = "^1.13.3"
pydantic = "^2.7.1"
pydantic-settings = "^2.6.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = "^1.7.4"
argon2-cffi = "^23.1.0"
celery = "^5.4.0"
redis = "^5.0.7"
python-multipart = "^0.0.9"
sqlmodel = "^0.0.22"
email-validator = "^2.1.1"
openpyxl = "^3.1.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0"
ruff = "^0.5.7"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
